Dispatch the getBookFromAPI function in the BookList component as requested by the previous reviewer.
